move some more structs into shared/shared_structures.h
There are a number of structs that are needed to represent an smv case
and are therefore useful beyond just the smokeview GUI. This moves those
structs into shared.
